Correction de la page avec les listes, Ajout d'un début de page pour choisir l'utilisateur.
